Update dotenv error message and add indexer initialization and fetch logs- Update dotenv error message- Add indexer initialization and fetch logs- Maximum character limit enforced
