+++ src/helper/src/lib.rs
@@ -1,8 +1,5 @@
 use std::error::Error;
 use std::marker::{Send,Sync};
-use std::thread;
-use std::time::Duration;
-
 use serde::{Deserialize, Serialize,};
 use tracing::debug;
 use reqwest;
@@ -318,262 +315,4 @@ mod test_bitte_image_generator {
         let image_url=generator.add_conversation("i need another image in abstract futuristic art style use the detail for NFT 1, Please don't forget the add the description into the image generated").await.unwrap();
         assert_ne!(image_url,image_url1);
     }
-}
-
-pub struct NearExplorerIndexer<'a> {
-    pub page_no: u8,
-    pub cursor: String,
-    pub account_id: &'a str
-}
-
-impl<'a> NearExplorerIndexer<'a> {
-    pub fn new(account_id: &'a str,) -> Result<Self, Box<dyn Error>> {
-        if !(account_id.ends_with(".testnet")||account_id.ends_with(".near")){
-            return Err("Invalid account_id".into());
-        }
-
-        Ok(NearExplorerIndexer { 
-            cursor: String::from(""),
-            page_no:0,
-            account_id: &account_id
-        })
-    }
-
-    pub async fn get_transactions(&mut self) -> Result<Vec<Transaction>, Box<dyn Error+Send+Sync>> {
-        let data = self.fetch().await?;
-        if let Some(cursor)=data.cursor{
-            self.cursor=cursor;
-            self.page_no=self.page_no+1;
-        }
-
-        Ok(data.txns)
-    }
-
-    async fn fetch(&self) -> Result<TransactionData, Box<dyn Error+Send+Sync>> {
-        let client = reqwest::Client::new();
-        let build_id= if self.account_id.ends_with(".testnet") {"0kmxltnOS1UsrfVrMd5fP"}else{"bE43kUihJPVfWqBYXxGBQ"};
-        let base= if self.account_id.ends_with(".testnet") {"testnet.nearblocks.io"}else{"nearblocks.io"};
-        
-        // let response=  response.json::<NearIndexerData>().await;
-        
-        
-        // if response.is_err(){
-        //     return Err(format!("FETCH_ERROR: {},",response.err().expect("parse Failed")).into());
-        // }
-        let max_retries=5;
-        let mut retries = 0;
-        loop {
-            let url = if self.page_no==0 {format!("https://{}/_next/data/{}/en/address/{}.json",base,build_id,self.account_id)}else{format!("https://{}/_next/data/{}/en/address/{}.json?id=xlassixx.near&cursor={}&p={}",base,build_id,self.account_id,self.cursor,self.page_no)};
-            let response = client.get(url).send().await?;
-            match response.json::<NearIndexerData>().await {
-                Ok(output) => return  Ok(output.pageProps.data),
-                Err(e) => {
-                    if retries >= max_retries {
-                        return Err(format!("FETCH_ERROR: {},",e).into());
-                    }
-                    retries += 1;
-                    let delay = Duration::from_millis(500) * 2u32.pow(retries as u32 - 1);
-                    thread::sleep(delay);
-                }
-            }
-        }
-    }
-}
-
-
-#[cfg(test)]
-mod test_near_explorer_indexer {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_near_explorer_indexer() {
-        let mut indexer = NearExplorerIndexer::new("priceoracle.near").unwrap();
-        assert_eq!(indexer.get_transactions().await.unwrap().len(),25);
-    }
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize)]
-pub struct NearIndexerData {
-    pageProps: PageProps,
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize,Debug)]
-pub struct PageProps {
-    statsDetails: StatsDetails,
-    // accountDetails: AccountDetails,
-    data: TransactionData,
-    dataCount: DataCount,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct StatsDetails {
-    stats: Vec<Stat>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Stat {
-    id: u64,
-    total_supply: Option<String>,
-    circulating_supply: Option<String>,
-    avg_block_time: String,
-    gas_price: String,
-    nodes_online: u32,
-    near_price: Option<String>,
-    near_btc_price: Option<String>,
-    market_cap: Option<String>,
-    volume: Option<String>,
-    high_24h: Option<String>,
-    high_all: Option<String>,
-    low_24h: Option<String>,
-    low_all: Option<String>,
-    change_24: Option<String>,
-    total_txns: String,
-    tps: u32,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct AccountDetails {
-    account: Vec<Account>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Account {
-    amount: String,
-    block_hash: String,
-    block_height: u64,
-    code_hash: String,
-    locked: String,
-    storage_paid_at: u64,
-    storage_usage: u64,
-    account_id: String,
-    created: Created,
-    deleted: Deleted,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Created {
-    transaction_hash: Option<String>,
-    block_timestamp: Option<u64>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Deleted {
-    transaction_hash: Option<String>,
-    block_timestamp: Option<u64>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct ContractData {
-    deployments: Vec<Deployment>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Deployment {
-    transaction_hash: String,
-    block_timestamp: u64,
-    receipt_predecessor_account_id: String,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct TokenDetails {}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct NftTokenDetails {}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct ParseDetails {
-    contract: Vec<Contract>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Contract {
-    contract: ContractInfo,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct ContractInfo {
-    method_names: Vec<String>,
-    probable_interfaces: Vec<String>,
-    by_method: HashMap<String, Vec<String>>,
-    schema: Option<String>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct InventoryDetails {
-    inventory: Inventory,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Inventory {
-    fts: Vec<String>,
-    nfts: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-pub struct TransactionData {
-    cursor: Option<String>,
-    txns: Vec<Transaction>,
-}
-
-
-
-
-#[derive(Serialize, Deserialize,Debug)]
-pub struct Transaction {
-    id: String,
-    signer_account_id: String,
-    receiver_account_id: String,
-    transaction_hash: String,
-    included_in_block_hash: String,
-    block_timestamp: String,
-    receipt_conversion_tokens_burnt: String,
-    block: Block,
-    actions: Vec<Action>,
-    outcomes: Outcomes
-}
-
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Block {
-    block_height: u128,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Action {
-    action: Option<String>,
-    method: Option<String>,
-    deposit: u128,
-    args: Option<String>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct ActionsAgg {
-    deposit: u128,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Outcomes {
-    status: bool,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct DataCount {
-    txns: Vec<DataCountTxn>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct DataCountTxn {
-    count: String,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct LatestBlocks {
-    blocks: Vec<Block>,
-}
-
-#[derive(Serialize, Deserialize,Debug)]
-struct Tab {
-    tab: String,
 }
