+++ src/near_indexer/.rustfmt.toml
@@ -1,0 +1,3 @@
+[tool.rustfmt]
+version = "nightly"
+
