+++ src/near_indexer/README.ME
@@ -1,0 +1,9 @@
+
+- Apply migrations
+```
+sea-orm-cli migrate up
+```
+
+```
+ sea-orm-cli generate entity -o migration/src/entity
+```
