+++ src/near_indexer/migration/src/entity/mod.rs
@@ -1,0 +1,5 @@
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.1
+
+pub mod prelude ;
+
+pub mod transaction ;
