+++ src/near_indexer/migration/src/entity/prelude.rs
@@ -1,0 +1,3 @@
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.1
+
+pub use super :: transaction :: Entity as Transaction ;
