+++ src/near_indexer/migration/src/entity/transaction.rs
@@ -1,0 +1,11 @@
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.1
+
+
+
+use sea_orm :: entity :: prelude :: * ;
+
+# [derive (Clone , Debug , PartialEq , DeriveEntityModel , Eq)] # [sea_orm (schema_name = "public" , table_name = "transaction")] pub struct Model { # [sea_orm (primary_key , auto_increment = false)] pub id : i32 , # [sea_orm (unique)] pub transaction_hash : String , pub signer_account_id : String , pub receiver_account_id : String , pub block_timestamp : String , pub block_height : i32 , pub action : String , pub method : String , pub outcomes_status : bool , pub tweet_id : String , pub image_url : String , pub user_to_notify : Option < String > , pub notarized_proof : Option < String > , pub zk_proof : Option < String > , }
+
+# [derive (Copy , Clone , Debug , EnumIter , DeriveRelation)] pub enum Relation { }
+
+impl ActiveModelBehavior for ActiveModel { }
